+++ packages/polaris-core/test/integration/tests/query-without-connection.test.ts
@@ -1,8 +1,9 @@
 import { PolarisServer } from '../../../src';
+import { allBooksNoConnectionData } from '../server-without-connection/schema/resolvers';
 import { startTestServer, stopTestServer } from '../server-without-connection/test-server';
 import { graphQLRequest } from '../server/utils/graphql-client';
-import * as argsQuery from './jsonRequestsAndHeaders/queryWithArgs.json';
-import * as simpleQuery from './jsonRequestsAndHeaders/simpleQueryWithoutConnection.json';
+import * as allBooksNoConnection from './jsonRequestsAndHeaders/allBooksNoConnection.json';
+import * as booksByTitle from './jsonRequestsAndHeaders/booksByTitle.json';
 
 let polarisServer: PolarisServer;
 
@@ -16,13 +17,14 @@ afterEach(async () => {
 
 describe('simple queries without connection', () => {
     it('all entities query', async () => {
-        const result: any = await graphQLRequest(simpleQuery.request, simpleQuery.headers);
-        expect(result.allBooks[0].title).toEqual('Book1');
-        expect(result.allBooks[1].title).toEqual('Book2');
+        const result: any = await graphQLRequest(allBooksNoConnection.request);
+        expect(result.allBooks[0].title).toEqual(allBooksNoConnectionData[0].title);
+        expect(result.allBooks[1].title).toEqual(allBooksNoConnectionData[1].title);
     });
 
     it('query with arguments', async () => {
-        const result: any = await graphQLRequest(argsQuery.request, argsQuery.headers);
-        expect(result.bookByTitle[0].title).toEqual('Book3');
+        const title = allBooksNoConnectionData[2].title;
+        const result: any = await graphQLRequest(booksByTitle.request, {}, { title });
+        expect(result.bookByTitle[0].title).toEqual(title);
     });
 });