+++ packages/polaris-core/test/integration/tests/snapshot-auto-max-1.test.ts
@@ -1,9 +1,9 @@
 import { PolarisServer } from '../../../src';
-import { initializeDatabase } from '../server/dal/data-initalizer';
 import { startTestServer, stopTestServer } from '../server/test-server';
-import { graphqlRawRequest } from '../server/utils/graphql-client';
+import { graphqlRawRequest, graphQLRequest } from '../server/utils/graphql-client';
 import { waitUntilSnapshotRequestIsDone } from '../server/utils/snapshot-client';
-import * as paginatedQuery from './jsonRequestsAndHeaders/paginatedQuery.json';
+import * as allBooksPaginated from './jsonRequestsAndHeaders/allBooksPaginated.json';
+import * as createBook from './jsonRequestsAndHeaders/createBook.json';
 
 let polarisServer: PolarisServer;
 beforeEach(async () => {
@@ -16,7 +16,8 @@ beforeEach(async () => {
             entitiesAmountPerFetch: 50,
         },
     });
-    await initializeDatabase();
+    await graphQLRequest(createBook.request, {}, { title: 'book01' });
+    await graphQLRequest(createBook.request, {}, { title: 'book02' });
 });
 afterEach(async () => {
     await stopTestServer(polarisServer);
@@ -26,8 +27,8 @@ describe('snapshot pagination tests with auto enabled', () => {
     describe('max page size is 1', () => {
         describe('snap request is false', () => {
             it('should paginated anyway', async () => {
-                const paginatedResult = await graphqlRawRequest(paginatedQuery.request, {
-                    ...paginatedQuery.headers,
+                const paginatedResult = await graphqlRawRequest(allBooksPaginated.request, {
+                    ...allBooksPaginated.headers,
                     'snap-request': false,
                 });
                 const pageIds = paginatedResult.extensions.snapResponse.pagesIds;
@@ -39,8 +40,8 @@ describe('snapshot pagination tests with auto enabled', () => {
             });
 
             it('should paginated according to minimal snap page size provided', async () => {
-                const paginatedResult = await graphqlRawRequest(paginatedQuery.request, {
-                    ...paginatedQuery.headers,
+                const paginatedResult = await graphqlRawRequest(allBooksPaginated.request, {
+                    ...allBooksPaginated.headers,
                     'snap-request': false,
                     'snap-page-size': 10,
                 });