+++ packages/polaris-core/test/integration/tests/subscription.test.ts
@@ -1,9 +1,11 @@
 import { PolarisServer, PolarisServerOptions } from '../../../src';
-import { initializeDatabase } from '../server/dal/data-initalizer';
 import * as polarisProperties from '../server/resources/polaris-properties.json';
 import { startTestServer, stopTestServer } from '../server/test-server';
 import { graphQLRequest } from '../server/utils/graphql-client';
 import { WebsocketClient } from '../server/utils/websocket-client';
+import * as createBook from './jsonRequestsAndHeaders/createBook.json';
+import * as subscriptionRequest from './jsonRequestsAndHeaders/subscriptionBookUpdated.json';
+import * as updateBooksByTitle from './jsonRequestsAndHeaders/updateBooksByTitle.json';
 
 const SUBSCRIPTION_ENDPOINT = `ws://localhost:${polarisProperties.port}/${polarisProperties.version}/subscription`;
 
@@ -16,7 +18,6 @@ beforeEach(async () => {
     };
     polarisServer = await startTestServer(subscriptionConfig);
     wsClient = new WebsocketClient(SUBSCRIPTION_ENDPOINT);
-    await initializeDatabase();
 });
 
 afterEach(async () => {
@@ -28,29 +29,10 @@ describe('subscription tests', () => {
     test('subscribing to book updates, and receiving a message once a book was updated', async () => {
         const title = 'Book1';
         const newTitle = 'Just a Title';
+        await graphQLRequest(createBook.request, {}, { title });
 
-        await wsClient.send(
-            `
-                subscription {
-                    bookUpdated {
-                        id
-                        title
-                    }
-                }
-            `,
-        );
-        await graphQLRequest(
-            `
-                mutation($title: String!, $newTitle: String!) {
-                   updateBooksByTitle(title: $title, newTitle: $newTitle) {
-                        id
-                        title
-                    }
-                }
-            `,
-            {},
-            { title, newTitle },
-        );
+        await wsClient.send(subscriptionRequest.request);
+        await graphQLRequest(updateBooksByTitle.request, {}, { title, newTitle });
 
         expect(wsClient.receivedMessages[0].bookUpdated.title).toBe(newTitle);
     });