+++ packages/polaris-core/test/integration/tests/warnings-enabled.test.ts
@@ -1,8 +1,7 @@
 import { PolarisServer, PolarisServerOptions } from '../../../src';
-import { initializeDatabase } from '../server/dal/data-initalizer';
 import { startTestServer, stopTestServer } from '../server/test-server';
 import { graphqlRawRequest } from '../server/utils/graphql-client';
-import * as booksWithWarnings from './jsonRequestsAndHeaders/queryForBooksWithWarnings.json';
+import * as booksWithWarnings from './jsonRequestsAndHeaders/allBooksWithWarnings.json';
 
 let polarisServer: PolarisServer;
 
@@ -13,7 +12,6 @@ describe('warnings enabled tests', () => {
                 shouldAddWarningsToExtensions: true,
             };
             polarisServer = await startTestServer(warningConfig);
-            await initializeDatabase();
         });
 
         afterEach(async () => {
@@ -21,18 +19,12 @@ describe('warnings enabled tests', () => {
         });
 
         it('warnings in the extensions of the response should be defined', async () => {
-            const result = await graphqlRawRequest(
-                booksWithWarnings.request,
-                booksWithWarnings.headers,
-            );
+            const result = await graphqlRawRequest(booksWithWarnings.request);
             expect(result.extensions.warnings).toBeDefined();
         });
 
         it('the relevant warnings should be returned in the extensions of the response', async () => {
-            const result = await graphqlRawRequest(
-                booksWithWarnings.request,
-                booksWithWarnings.headers,
-            );
+            const result = await graphqlRawRequest(booksWithWarnings.request);
             expect(result.extensions.warnings.length).toBe(2);
             expect(result.extensions.warnings[0]).toEqual('warning 1');
             expect(result.extensions.warnings[1]).toEqual('warning 2');