+++ packages/polaris-typeorm/src/models/common-model.ts
@@ -18,13 +18,13 @@ export abstract class CommonModel {
     @Column({ nullable: true })
     protected createdBy?: string;
 
-    @CreateDateColumn()
+    @Column({ default: 'NOW()' })
     protected creationTime: Date;
 
     @Column({ nullable: true })
     protected lastUpdatedBy?: string;
 
-    @UpdateDateColumn()
+    @UpdateDateColumn({ onUpdate: 'NOW()' })
     protected lastUpdateTime: Date;
 
     @Column()